Move table creator to core

+++ pkg/client/client.go
@@ -3,6 +3,7 @@ package client
 import (
 	"context"
 	"encoding/json"
+	"github.com/cloudquery/cloudquery/pkg/database"
 	"io"
 	"os"
 	"path/filepath"
@@ -103,6 +104,8 @@ type Client struct {
 	NoVerify bool
 	// Optional: DSN connection information for database client will connect to
 	DSN string
+	// Optional: Skips Building tables on fetch execution
+	SkipBuildTables bool
 	// Optional: HubProgressUpdater allows the client creator to get called back on download progress and completion.
 	HubProgressUpdater ui.Progress
 	// Optional: Logger framework can use to log.
@@ -113,6 +116,8 @@ type Client struct {
 	Hub registry.Hub
 	// manager manages all plugins lifecycle
 	Manager *plugin.Manager
+	// TableCreator defines how table are created in the database
+	TableCreator database.TableCreator
 	// pool is a list of connection that are used for policy/query execution
 	pool *pgxpool.Pool
 }
@@ -123,6 +128,7 @@ func New(ctx context.Context, options ...Option) (*Client, error) {
 		PluginDirectory:    filepath.Join(".", ".cq", "providers"),
 		PolicyDirectory:    ".",
 		NoVerify:           false,
+		SkipBuildTables:    false,
 		HubProgressUpdater: nil,
 		RegistryURL:        registry.CloudQueryRegistryURl,
 		Logger:             logging.NewZHcLog(&zerolog.Logger, ""),
@@ -136,7 +142,9 @@ func New(ctx context.Context, options ...Option) (*Client, error) {
 	if err != nil {
 		return nil, err
 	}
-
+	if c.TableCreator == nil {
+		c.TableCreator = database.NewBaseTableCreator(c.Logger)
+	}
 	poolCfg, err := pgxpool.ParseConfig(c.DSN)
 	if err != nil {
 		return nil, err
@@ -156,6 +164,13 @@ func (c *Client) DownloadProviders(ctx context.Context) error {
 }
 
 func (c *Client) Fetch(ctx context.Context, request FetchRequest) error {
+	if !c.SkipBuildTables {
+		for _, provider := range request.Providers {
+			if err := c.BuildProviderTables(ctx, provider.Name); err != nil {
+				return err
+			}
+		}
+	}
 	errGroup, gctx := errgroup.WithContext(ctx)
 	for _, providerConfig := range request.Providers {
 		providerConfig := providerConfig
@@ -232,6 +247,10 @@ func (c Client) GetProviderSchema(ctx context.Context, providerName string) (*cq
 		return nil, err
 	}
 	defer func() {
+		if providerPlugin.Version() == plugin.Unmanaged {
+			c.Logger.Warn("Not closing unmanaged provider", "provider", providerName)
+			return
+		}
 		if err := c.Manager.KillProvider(providerName); err != nil {
 			c.Logger.Warn("failed to kill provider", "provider", providerName)
 		}
@@ -246,11 +265,34 @@ func (c Client) GetProviderConfiguration(ctx context.Context, providerName strin
 		return nil, err
 	}
 	defer func() {
+		if providerPlugin.Version() == plugin.Unmanaged {
+			c.Logger.Warn("Not closing unmanaged provider", "provider", providerName)
+			return
+		}
 		if err := c.Manager.KillProvider(providerName); err != nil {
-			c.Logger.Warn("failed to kill provider", "provider", providerName)
+			c.Logger.Warn("failed to close provider", "provider", providerName)
 		}
 	}()
 	return providerPlugin.Provider().GetProviderConfig(ctx, &cqproto.GetProviderConfigRequest{})
+}
+
+func (c *Client) BuildProviderTables(ctx context.Context, providerName string) error {
+	schema, err := c.GetProviderSchema(ctx, providerName)
+	if err != nil {
+		return err
+	}
+	conn, err := c.pool.Acquire(ctx)
+	if err != nil {
+		return err
+	}
+	defer conn.Release()
+	for name, t := range schema.ResourceTables {
+		c.Logger.Info("creating tables for resource for provider", "resource_name", name, "provider", schema.Name, "version", schema.Version)
+		if err := c.TableCreator.CreateTable(ctx, conn, t, nil); err != nil {
+			return err
+		}
+	}
+	return nil
 }
 
 func (c Client) DownloadPolicy(ctx context.Context, args []string) error {
